requirements.txt
psutil
python-telegram-bot>=21.10